upgrading to nexJS13, using server component and client components and app directory
