+++ app/about/page.js
@@ -1,0 +1,12 @@
+import '../globals.scss';
+export default function About() {
+  return (
+    <>
+      <main className='main'>
+        <h2 className='logoHome'>MEMOMO </h2>
+        <h1>A propos</h1>
+        <p>blabla</p>
+      </main>
+    </>
+  );
+}
