+++ app/globals.scss
@@ -1,5 +1,25 @@
+/* -----------IMPORTS --------------*/
 @import '/styles/colorsVariables.scss';
+@import '/styles/reset.scss';
 @import '/styles/globalClasses.scss';
+/* -----------GENERAL LAYOUT ON ALL PAGES --------------*/
+
+body {
+  background-color: $main-color-dark;
+  color: $light-color-white;
+  font-family: var(--font-rounded-elegance);
+  border: 2px solid pink;
+}
+
+a:hover {
+  cursor: pointer;
+}
+a {
+  text-decoration: none;
+}
+button {
+  background: transparent;
+}
 .mainWrapper {
   border: 3px solid $third-color-darker;
   display: flex;
@@ -18,7 +38,6 @@
   max-width: 750px;
   margin: 0 auto;
   height: 100%;
-
   display: flex;
   flex-direction: column;
   justify-content: center;
