+++ components/memoCard/index.js
@@ -1,9 +1,10 @@
+'use client';
 import { useEffect, useState } from 'react';
 import Image from 'next/image';
 import styles from '../../styles/MemoCard.module.scss';
-import { putMemoRequest } from '../../requests/memos';
+import { putMemoRequest } from '../../lib/requestsDatas';
 
-export default function Memocard({ id, content, colour, boardId, deleteMemo }) {
+export default function Memocard({ id, content, colour, boardId }) {
   const [colourClass, setColourClass] = useState('');
   const [memoContent, setMemoContent] = useState(content);
   const fillMemo = (e) => {
@@ -28,6 +29,7 @@ export default function Memocard({ id, content, colour, boardId, deleteMemo }) {
 
   useEffect(() => {
     setColourClass(`colour${colour}`);
+    console.log(memoContent);
   }, []);
 
   return (
@@ -42,7 +44,7 @@ export default function Memocard({ id, content, colour, boardId, deleteMemo }) {
           value={memoContent}
         />
       </div>
-      <div className={styles.memoImg}>
+      {/* <div className={styles.memoImg}>
         <button
           onClick={() => {
             deleteMemo(id);
@@ -55,7 +57,7 @@ export default function Memocard({ id, content, colour, boardId, deleteMemo }) {
             alt='X to delete the memo'
           />
         </button>
-      </div>
+      </div> */}
     </div>
   );
 }
