+++ config/site.js
@@ -1,0 +1,10 @@
+export const siteConfig = {
+  name: 'MEMOMO',
+  description: 'Création de tableaux personnels contenant des listes',
+
+  ogImage: 'https://tx.shadcn.com/og.jpg',
+  links: {
+    githubFront: 'https://github.com/ilalande/Memomo',
+    githubBack: 'https://github.com/ilalande/memomoBack',
+  },
+};
