--- pages/about.js
@@ -1,11 +1,0 @@
-import PrimaryLayout from '../components/primaryLayout';
-
-export default function About() {
-  return (
-    <>
-      <PrimaryLayout>
-        <p>blabla</p>
-      </PrimaryLayout>
-    </>
-  );
-}
