--- styles/About.module.scss
@@ -1,14 +1,0 @@
-@import '/styles/colorsVariables.scss';
-@import '/styles/globalClasses.scss';
-.mainWrapper {
-  border: 3px solid $third-color-darker;
-  display: flex;
-  flex-direction: column-reverse;
-  width: 100%;
-  height: 100vh;
-  max-width: 1250px;
-  font-family: var(--font-rounded-elegance);
-  /* To fix menu at left size */
-  align-items: flex-start;
-  margin: 0 auto;
-}
