+++ styles/Board.module.scss
@@ -1,7 +1,11 @@
 /* HEADER : title and add buttons to add memos */
 @import '/styles/colorsVariables.scss';
 @import '/styles/globalClasses.scss';
-
+.main {
+  display: block;
+  height: 100%;
+  margin: 0 auto;
+}
 .boardHeader {
   height: 15%;
 }
@@ -18,31 +22,6 @@
   font-family: var(--font-tondu);
   color: $second-color-lighter;
   padding: 1rem;
-}
-.plusButtonOnBoard {
-  font-size: 1rem;
-  margin: 1rem;
-}
-
-.plusButtonOnBoard button:hover {
-  cursor: pointer;
-}
-
-.addMemo2 {
-  @extend .plusButton;
-  border-color: $memo-2;
-  color: $memo-2;
-  /* background-color: red; */
-}
-.addMemo1 {
-  @extend .plusButton;
-  border-color: $memo-1;
-  color: $memo-1;
-}
-.addMemo3 {
-  @extend .plusButton;
-  border-color: $memo-3;
-  color: $memo-3;
 }
 /* BOARD AREA : existing memod znd possibility to handle them */
 .boardarea {
