--- styles/globals.scss
@@ -1,21 +1,0 @@
-/* -----------IMPORTS --------------*/
-@import '/styles/colorsVariables.scss';
-@import '/styles/reset.scss';
-
-/* -----------GENERAL LAYOUT ON ALL PAGES --------------*/
-
-body {
-  background-color: $main-color-dark;
-  color: $light-color-white;
-  font-family: var(--font-rounded-elegance);
-}
-
-a:hover {
-  cursor: pointer;
-}
-a {
-  text-decoration: none;
-}
-button {
-  background: transparent;
-}
